requirements.txt
selenium
webdriver_manager
requests
beautifulsoup4
pytest-html